requirements.txt
amqtt==0.11.3
paho-mqtt==2.1.0
PySide6==6.9.0
